Modifications to source code, new GUI style and transfer function generator
A lot of changes in this commit.

The source code was modified to stop QmitkStdMultiWidget and QmitkRenderWindowWidget from doing global modifications in the mitk::RenderWindowManager and at the mitk::StandaloneDataSotrage which is given to them, those changes may break some code.

The GUI received a new style, but unexpectedly the svg needed to be disable to not crash the code at the app stratup. NEEDS FIX.

Also, MITK's function generator plugin was incorporated to the GUI, all berry related stuff was removed making the required substitutions and other modifications.

Finally the current devolpment is occuring inside `experiments` folder in FileLoading application. A singleton class to manage DataStorage was created, some workaround was done to retrieve all nodes as a list that can be read from singleton class for DataStorage. With this last feature was possible to create a DataManagerView that can hide/show specific nodes loaded by the application. Those changes need to be translated to the original GUI source code, requiring a lot of modifications.