Improvements in the modular structure of the code. SVG plugin fixed. MovieMaker.
Code is more modular now. The modules can be dependent on each other, I believe CMakeLists.txt can be written in a clear way to do this task.

The SVG Qt plugin crashed because I did an upgrade in the OS, the upgrade changed the svg plugin, to fix, I uninstalled the qt version for android and desktop, reinstalled again and fixed the `/usr/lib/x86_64-linux-gnu/libQt5Svg.so.5`. There is a version of this file inside Qt 5.12.8 folder, I don't understand why the system lib is used instead.

MovieMaker funcionalities were imported to the application, now I will customize it to have a distinct look at the application and not reseamble the original so much.